testing update
- scaffolding now meets the documentation of 1:1 mapping to main
  src/botmation directory
